Make sure AutoGeneratedDataFolder is instantiated at all times. Fixes ITs on new computers.

+++ Testing/ArgumentProcessingServiceIT.py
@@ -18,6 +18,8 @@ class ArgumentProcessingServiceIT(unittest.TestCase):
     def tearDown(self):
         if self.current_working_dir != "/":
             for file in os.listdir(self.current_working_dir + "/" + RandomizedDataGenerator.GENERATED_DATA_FOLDER):
+                if file == "__init__.py":
+                    continue
                 os.remove(self.current_working_dir + "/" + RandomizedDataGenerator.GENERATED_DATA_FOLDER + "/" + file)
 
     def testInputTextFileCorrectlyParsed(self):
